# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Vespin Retro Companion API
  version: 0.1.0
  description: |
    HTTP API for the Vespin Retro smart speaker companion mobile app.

    **HCI course project — hardware interaction is fully simulated.**
    No real Bluetooth, audio processing, firmware OTA, or device connectivity.
    Simulated fields (firmware_version, battery_level, is_connected) are seeded
    on the backend at device creation time.

    ## Authentication
    JWT bearer tokens. Two roles: `guest` and `registered`. Guests can pair
    devices, manage EQ profiles, and run party sessions. Preferences and
    firmware check are registered-only.

    ## Conventions
    - JSON request/response bodies, camelCase field names.
    - Timestamps are RFC 3339 strings in UTC.
    - IDs are UUIDs.
    - PATCH for partial updates (never PUT).
    - Error envelope: `{ "error": { "code": "...", "message": "..." } }`.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.example.com
    description: Production (placeholder — replace with real domain)
security:
  - bearerAuth: []
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      tags: [Users]
      summary: Update current user profile
      description: |
        Currently only `displayName` is editable. Available to both guests and
        registered users — guests can personalize their displayed name before
        committing to a full account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/preferences:
    get:
      tags: [Users]
      summary: Get current user preferences
      description: |
        Registered users only. Returns default values when the user has
        never set a preference (no row exists yet).
      responses:
        '200':
          description: Current preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preferences'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/GuestForbidden'

    patch:
      tags: [Users]
      summary: Update current user preferences
      description: |
        Registered users only. Upserts: creates the preferences row with
        defaults plus the patch values if it doesn't exist, otherwise updates
        only the patched fields.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreferencesUpdateRequest'
      responses:
        '200':
          description: Updated preferences (full current state)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preferences'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/GuestForbidden'

  # ---------------------------------------------------------------------------
  # Devices
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---------------------------------------------------------------------------
  # Shared parameters
  # ---------------------------------------------------------------------------
...
  responses:
    ValidationFailed:
      description: Request body or query parameters failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: validation_failed
              message: Request validation failed.
              details:
                fields:
                  email: must be a valid email address
                  password: must be at least 8 characters
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: unauthorized
              message: Authentication required.
    GuestForbidden:
      description: Endpoint requires a registered account
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: guest_endpoint_forbidden
              message: This endpoint is only available to registered users.
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code (snake_case).
              example: not_found
            message:
              type: string
              description: Human-readable error message.
            details:
              type: object
              additionalProperties: true
              description: |
                Optional structured details. For validation errors, contains
                a `fields` object mapping field names to per-field error
                messages.

    # --- User ----------------------------------------------------------------
    User:
      type: object
      required: [id, role, createdAt]
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [guest, registered]
        email:
          type: [string, "null"]
          format: email
          description: Null for guest users.
        displayName:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
        convertedAt:
          type: [string, "null"]
          format: date-time
          description: Set when a guest converts to a registered account.
    UserUpdateRequest:
      type: object
      properties:
        displayName:
          type: string

    # --- Preferences ---------------------------------------------------------
    Preferences:
      type: object
      required: [theme, language, notificationsEnabled]
      properties:
        theme:
          type: string
          enum: [light, dark, system]
        language:
          type: string
          description: ISO 639-1 code.
          example: en
        notificationsEnabled:
          type: boolean
    PreferencesUpdateRequest:
      type: object
      properties:
        theme:
          type: string
          enum: [light, dark, system]
        language:
          type: string
        notificationsEnabled:
          type: boolean

    # --- Auth DTOs -----------------------------------------------------------
...